WebDevLab02/requirements.txt
streamlit
requests
plotly.express